pyproject.toml
[tool.poetry]
name = "agent-receptionist"
version = "0.1.0"
description = "AI-powered voice receptionist for dental offices"
authors = []
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"

# Web framework
fastapi = ">=0.115.0"
uvicorn = {extras = ["standard"], version = ">=0.34.0"}

# Database
sqlalchemy = {extras = ["asyncio"], version = ">=2.0.0"}
alembic = ">=1.14.0"
aiosqlite = ">=0.20.0"

# Validation & config
pydantic = ">=2.0.0"
pydantic-settings = ">=2.0.0"
python-dotenv = ">=1.0.0"

# HTTP client
httpx = ">=0.28.0"

# Voice pipeline
deepgram-sdk = ">=3.0.0"
cartesia = ">=1.0.0"
twilio = ">=9.0.0"
websockets = ">=12.0"
audioop-lts = {version = ">=0.2.1", python = ">=3.13"}

# Agent framework
langgraph = ">=0.4.0"
langchain-openai = ">=0.3.0"
openai = ">=1.0.0"
greenlet = "^3.3.2"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.24.0"
ruff = ">=0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 120
